+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `Carousel`: Add `shadowSize` prop to handle cards with shadows.
+
 ## [10.7.0] - 2022-05-09
 
 Deprecations/Changes: